优化了代码结构，增加了菜单功能。 问题： - 刷机完毕后选择界面没反应 - 无设置 - 不能自由管理选择文件
